XML_search/requirements.txt
# Telegram Bot API
python-telegram-bot==20.7

# Работа с конфигурацией
python-dotenv==1.0.0

# XML обработка
lxml==4.9.3

# База данных
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Геометрические операции
shapely==2.0.2

# Дополнения от ошибок
numpy==1.24.3
unidecode==1.3.6

# Логирование и утилиты
pytz==2024.1
typing-extensions>=4.5.0

# Опциональные (но рекомендуемые) пакеты для улучшенной производительности
ujson>=5.8.0
aiohttp==3.8.5

# Дополнительные пакеты
pyproj==3.6.1
pandas==2.0.3
requests==2.31.0
asyncio==3.4.3

# Добавляю библиотеку для транслитерации
httpx==0.25.2
transliterate==1.10.2

# Добавляю библиотеку для нечеткого поиска
python-Levenshtein==0.25.0

# Добавляю библиотеку для работы с Windows API
pywin32==306